Split up root
